refactor: moves logger related files to new logger folder.
