+++ src/blueprints/tv2/tv2-logger-facade.ts
@@ -1,4 +1,4 @@
-import { Logger } from '../../logger'
+import { Logger } from '../../logger/logger'
 import { Tv2ConsoleLogger } from './tv2-console-logger'
 
 export class Tv2LoggerFacade {
