+++ src/presentation/facades/event-server-facade.ts
@@ -1,7 +1,7 @@
 import { EventServer } from '../emitters/interfaces/event-server'
 import { WebSocketEventServer } from '../emitters/web-socket-event-server'
 import { EventEmitterFacade } from './event-emitter-facade'
-import { LoggerFacade } from '../../logger-facade'
+import { LoggerFacade } from '../../logger/logger-facade'
 
 export class EventServerFacade {
   public static createEventServer(): EventServer {
