+++ src/presentation/index.ts
@@ -4,8 +4,8 @@ import { BaseController } from './controllers/base-controller'
 import { ControllerFacade } from './facades/controller-facade'
 import { EventServerFacade } from './facades/event-server-facade'
 import { ServiceFacade } from '../business-logic/facades/service-facade'
-import { Logger } from '../logger'
-import { LoggerFacade } from '../logger-facade'
+import { Logger } from '../logger/logger'
+import { LoggerFacade } from '../logger/logger-facade'
 
 export * from './controllers/rundown-controller'
